[refactor] 컴포넌트 파일 위치변경
